refactor: moves cd and ls functions into a separate module navigation.js
